+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/activity/icon/IconRepository.kt
@@ -13,7 +13,7 @@ package io.github.pandier.intellijdiscordrp.activity.icon
  * When these changes make it into the repository, the icons
  * will start displaying correctly.
  */
-private const val ICON_REPOSITORY_COMMIT = "bf29ce96ea6b2d6a1af8a1c064b9fd95be5f7011"
+private const val ICON_REPOSITORY_COMMIT = "3fc1e26890027e9e32dfaba2e6157d8a3cb731bc"
 const val ICON_REPOSITORY =
     "https://raw.githubusercontent.com/pandier/intellij-discord-rp/$ICON_REPOSITORY_COMMIT/icons"
 const val FILE_ICON_REPOSITORY = "$ICON_REPOSITORY/file"